variable functions in separate header
